♻️ refactor(tokens): use dedicated module for token calculations
- move token-related functions to utils.tokens module for better organization
- update imports across various files to reflect new module structure
