+++ dev_scripts/test_tiktoken.py
@@ -1,4 +1,4 @@
-from argoproxy.utils import get_tiktoken_encoding_model
+from argoproxy.utils.tokens import get_tiktoken_encoding_model
 
 if __name__ == "__main__":
     from argoproxy.models import CHAT_MODELS, EMBED_MODELS
