+++ src/argoproxy/endpoints/chat.py
@@ -20,10 +20,9 @@ from ..types import (
     NonStreamChoice,
     StreamChoice,
 )
+from ..utils.tokens import calculate_prompt_tokens, count_tokens
 from ..utils.transport import send_off_sse
 from ..utils.utils import (
-    calculate_prompt_tokens,
-    count_tokens,
     make_bar,
     resolve_model_name,
 )
