+++ src/argoproxy/endpoints/responses.py
@@ -26,10 +26,9 @@ from ..types import (
     ResponseTextDoneEvent,
     ResponseUsage,
 )
+from ..utils.tokens import calculate_prompt_tokens, count_tokens
 from ..utils.transport import send_off_sse
 from ..utils.utils import (
-    calculate_prompt_tokens,
-    count_tokens,
     make_bar,
     resolve_model_name,
 )
